wic: restore selinux label for fstab
Issue: LINUXEXEC-26357

When selinux is enabled, the selinux labels are preserved for all files
in the wic image except /etc/fstab. This is because the original fstab
was replaced by a newly created fstab. Its extended attributes are not
preserved. Restore the new fstab's selinux label to ensure all files are
properly labeled with selinux context at build time. This avoids
relabeling selinux context on first boot.

(LOCAL REV: NOT UPSTREAM) -- Will send to upstream later

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>