sysvinit: backport patch for fixing one issue of pidof
Issue: LIN1024-5356

When a program is a link to an executable, and this executable is
handled by update-alternatives, renamed to a different name, cmd
`pidof full_path_of_program` will not work.

Eg:
/usr/sbin/httpd -> /usr/sbin/httpd.apache2
`pidof /usr/sbin/httpd` cannot get pid of the process httpd

Backport the patch to fix above issue.

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20241123

Signed-off-by: Changqing Li <changqing.li@windriver.com>