rsync: fix CVE-2024-12084
Issue: LIN1024-5751

A heap-based buffer overflow flaw was found in the rsync daemon. This issue is due
to improper handling of attacker-controlled checksum lengths (s2length) in the code.
When MAX_DIGEST_LEN exceeds the fixed SUM_LENGTH (16 bytes), an attacker can write
out of bounds in the sum2 buffer.

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20250116

Signed-off-by: Archana Polampalli <archana.polampalli@windriver.com>