rsync: fix CVE-2024-12085
Issue: LIN1024-5752

A flaw was found in the rsync daemon which could be triggered when rsync compares
file checksums. This flaw allows an attacker to manipulate the checksum length
(s2length) to cause a comparison between a checksum and uninitialized memory and
leak one byte of uninitialized stack data at a time.

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20250116

Signed-off-by: Archana Polampalli <archana.polampalli@windriver.com>