rsync: fix CVE-2024-12087
Issue: LIN1024-5754

A path traversal vulnerability exists in rsync. It stems from behavior enabled
by the `--inc-recursive` option, a default-enabled option for many client options
and can be enabled by the server even if not explicitly enabled by the client.
When using the `--inc-recursive` option, a lack of proper symlink verification
coupled with deduplication checks occurring on a per-file-list basis could allow
a server to write files outside of the client's intended destination directory.
A malicious server could write malicious files to arbitrary locations named after
valid directories/paths on the client.

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20250116

Signed-off-by: Archana Polampalli <archana.polampalli@windriver.com>