rsync: fix CVE-2024-12088
Issue: LIN1024-5755

A flaw was found in rsync. When using the `--safe-links` option, rsync fails to
properly verify if a symbolic link destination contains another symbolic link within it.
This results in a path traversal vulnerability, which may lead to arbitrary file write
outside the desired directory.

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20250116

Signed-off-by: Archana Polampalli <archana.polampalli@windriver.com>