rsync: fix CVE-2024-12747
Issue: LIN1024-6409

A flaw was found in rsync. This vulnerability arises from a race condition during
rsync's handling of symbolic links. Rsync's default behavior when encountering
symbolic links is to skip them. If an attacker replaced a regular file with a
symbolic link at the right time, it was possible to bypass the default behavior
and traverse symbolic links. Depending on the privileges of the rsync process,
an attacker could leak sensitive information, potentially leading to privilege escalation.

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20250116

Signed-off-by: Archana Polampalli <archana.polampalli@windriver.com>