minicom/wayland: correct the SRC_URIs
Issue: LIN1024-10068

For minicom, the old SRC_URI is not available and need update.
For wayland, wayland-utils and wayland-protocals, not only
the SRC_URIs are updated, but also the tar.xz are no longer
available. We have replaced them with the tar.bz2 barballs.

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20250702

Signed-off-by: Guocai He <guocai.he.cn@windriver.com>